refactor: move usage metric handler into dedicated package
